change long id to int

+++ src/main/java/com/example/repositories/ConfirmationTokenRepository.java
@@ -8,6 +8,6 @@ public interface ConfirmationTokenRepository
 {
     public Optional<ConfirmationToken> findByToken(String token);
     public void confirmToken(ConfirmationToken token);
-    public Long maxId();
+    public int maxId();
     public void save(ConfirmationToken token);
 }
